+++ CHANGELOG.md
@@ -1,51 +1,70 @@
+### Changes: V2.5.0
+
+- Added Shadowlands reputations
+- Added ability to colour the reputations by standing (Friendly, Honoured, etc.)
+
 ### Changes: V2.4.0
+
 - Added more reputations to the addon (They're currently hidden but can be shown in the settings)
 - Added new tooltip to header to give information about all characters
 - Updated scrollbar styling to make it stand out more
 
 ### Changes: V2.3.0
+
 - Added icon to reputation panel of character window (Can be hidden in the settings)
 - Added ability to add a custom sort order to your toons
 - Added tooltip to icons to give more info
 
 ### Changes: V2.2.0
+
 - Added ability to filter characters by server or connected
 - Added ability to group characters by server or connected realm
 - Fixed bug with libs
 
 ### Changes: V2.1.3
+
 - Removed chat logging
 
 ### Changes: V2.1.2
+
 - Fixed character scrolling bug
 
 ### Changes: V2.1.0
+
 - Swapped gold tracking for an estimation approach due to issues with the accuracy and only tracking the gold earned while the addon is enabled
 - Added functionality to forget characters that you no longer wish to track
 
 ### Changes: V2.0.3, V2.0.4
+
 - Fixed libs v2
 
 ### Changes: V2.0.2
+
 - Fixed frame template inheritence
 
 ### Changes: V2.0.1
+
 - Fixed libs
 
 ### Changes: V2.0.0
+
 - Updated for patch 9.0.1
 
 ### Changes: V1.2.1
+
 - Reordered Libs
 
 ### Changes: V1.2.0
+
 - Removed font size config as redundant due to bugs and adding scrolling
 - Made the character you're on show as the first character in the list
 
 ### Changes: V1.1.1
+
 - Fixed styling of scrollbar when using ElvUI
 - Fixed issue with scrollbar when changing the max number of visible characters while keeping the window open
 
 ### Changes: V1.1.0
+
 - Added ability to change font size
 - Added horizontal scrolling